bunch of changes I forgot to commit individually

+++ cron/1.characters.nolock.php
@@ -4,27 +4,39 @@ use cvweiss\redistools\RedisTimeQueue;
 
 require_once "../init.php";
 
+if ($redis->get("zkb:reinforced") == true) exit();
 $esi = new RedisTimeQueue('tqApiESI', 3600);
+$esiSkipped = new RedisTimeQueue('tqApiESI-skipped', 3600);
 if (date('i') == 22 || $esi->size() < 100) {
     $esis = $mdb->find("scopes", ['scope' => 'esi-killmails.read_killmails.v1']);
     foreach ($esis as $row) {
         $charID = $row['characterID'];
         $esi->add($charID);
     }
+    Util::out("Skipped " . $esiSkipped->size() . " characters in the last hour");
 }
-
 if ($esi->size() == 0) exit();
-$perSecond = max(ceil($esi->size() / 3600) - 3, 1);
-$sleepTime = min(10000, max(100, floor(10000 / $perSecond)));
-$guzzler = new Guzzler($perSecond, $sleepTime);
+
+$guzzler = new Guzzler(20, 1000);
 
 $minute = date('Hi');
 while ($minute == date('Hi')) {
+    Status::checkStatus($guzzler, 'sso');
     Status::checkStatus($guzzler, 'esi');
     $charID = (int) $esi->next(false);
     if ($charID > 0) {
         $row = $mdb->findDoc("scopes", ['characterID' => $charID, 'scope' => "esi-killmails.read_killmails.v1"], ['lastFetch' => 1]);
         if ($row != null) {
+            // Check to see if char's corp has an api, if they do, skip them unless the char hasn't been fetched before (look at maxKillID)
+            $corpID = Info::getInfoField("characterID", $charID, "corporationID");
+            $hasCorp = $mdb->count("scopes", ['corporationID' => $corpID, 'scope' => "esi-killmails.read_corporation_killmails.v1"]);
+            if ($hasCorp > 0 && $corpID > 0 && isset($row['maxKillID'])) {
+                //$corpName = Info::getInfoField("corporationID", $corpID, "name");
+                //Util::out("Skipping $charID for $corpName");
+                $esiSkipped->add($charID);
+                continue;
+            }
+
             $params = ['row' => $row, 'esi' => $esi];
             $refreshToken = $row['refreshToken'];
 
@@ -95,12 +107,13 @@ function success($guzzler, $params, $content)
         $successes = (int) @$row['successes'];
         $successes++;
         $mdb->set("scopes", $row, ['maxKillID' => $maxKillID, 'corporationID' => $corpID, 'lastFetch' => $mdb->now(), 'errorCount' => 0, 'successes' => $successes]);
+        $mdb->set("scopes", ['characterID' => $charID], ['corporationID' => $corpID]);
         $redis->setex("apiVerified:$charID", 86400, time());
 
         // Check active chars once an hour, check inactive chars less often
         $esi = new RedisTimeQueue('tqApiESI', 3600);
         if ($maxKillID > ($redis->get('zkb:topKillID') - 1000000)) {
-            $numHours = rand(24,48);
+            $numHours = rand(3, 4);
             $esi->setTime($charID, time() + (3600 * $numHours));
         }
 
@@ -137,19 +150,22 @@ function fail($guzzer, $params, $ex)
     $row = $params['row'];
     $esi = $params['esi'];
     $charID = $row['characterID'];
-    $name = Info::getInfoField("characterID", $charID, "name");
     $code = $ex->getCode();
+
+    $json = json_decode($params['content'], true);
+    $code = isset($json['sso_status']) ? $json['sso_status'] : $code;
 
     switch ($code) {
         case 400: // Server timed out during SSO authentication
-        case 403: // Server decided to throw a 403 during SSO authentication when that throws a 502...
+        case 420:
         case 500:
         case 502: // Server error, try again in 5 minutes
         case 503:
             $esi->setTime($charID, time() + 300);
             break;
         default:
-            echo "killmail: " . $ex->getMessage() . "\nkillmail content: " . $params['content'] . "\n";
+        case 403: // Server decided to throw a 403 during SSO authentication when that throws a 502...
+            echo "killmail char $charID: " . $ex->getMessage() . "\nkillmail content: " . $params['content'] . "\n";
     }
 }
 
@@ -162,17 +178,20 @@ function accessTokenFail(&$guzzler, &$params, $ex)
     $charID = $row['characterID'];
     $code = $ex->getCode();
 
+    $json = json_decode($params['content'], true);
+    if (@$json['error'] == 'invalid_grant') {
+        $mdb->remove("scopes", $row);
+        $esi->remove($charID);
+        return;
+    }
+
     switch ($code) {
-        case 400:
-        case 403: // No permission
-            $mdb->remove("scopes", $row);
-            $esi->remove($charID);
-            break;
         case 500:
         case 502: // Server error, try again in 5 minutes
             $esi->setTime($charID, time() + 300);
             break;
         default:
-            echo "token: " . $ex->getMessage() . "\n";
+            Util::out("char token: $charID " . $ex->getMessage() . "\n\n" . $params['content']);
     }
+    $esi->setTime($charID, time() + rand(1, 7200));
 }
